+++ Src/QueryCompiler.cs
@@ -30,6 +30,24 @@ namespace Mdbc
 	public static class QueryCompiler
 	{
 		#region Operators
+		static Expression ExistsExpression(Expression field, BsonValue args)
+		{
+			return Expression.Equal(
+				Expression.Call(Data, typeof(BsonDocument).GetMethod("Contains"), field),
+				Expression.Constant(LanguagePrimitives.IsTrue(Actor.ToObject(args)), typeof(bool)));
+		}
+		static Expression NotExpression(Expression field, BsonValue args)
+		{
+			switch (args.BsonType)
+			{
+				case BsonType.Document:
+					return Expression.Not(FieldExpression(field, args.AsBsonDocument));
+				case BsonType.RegularExpression:
+					return Expression.Not(MatchesExpression(field, args));
+				default:
+					throw new ArgumentException("Invalid form of $not.");
+			}
+		}
 		static bool EQ(BsonDocument document, string name, BsonValue value)
 		{
 			bool ok = false;
@@ -41,16 +59,9 @@ namespace Mdbc
 			}
 			return ok ? false : value.BsonType == BsonType.Null;
 		}
-		static bool NE(BsonDocument document, string name, BsonValue value)
+		static Expression GTExpression(Expression field, BsonValue args)
 		{
-			bool ok = false;
-			foreach (var data in document.GetNestedValues(name))
-			{
-				ok = true;
-				if (data.CompareTo(value) != 0)
-					return true;
-			}
-			return ok ? false : value.BsonType != BsonType.Null;
+			return Expression.Call(GetMethod("GT"), Data, field, Expression.Constant(args, typeof(BsonValue)));
 		}
 		static bool GT(BsonDocument document, string name, BsonValue value)
 		{
@@ -64,6 +75,10 @@ namespace Mdbc
 
 			return ok ? false : value.BsonType == BsonType.Null;
 		}
+		static Expression GTEExpression(Expression field, BsonValue args)
+		{
+			return Expression.Call(GetMethod("GTE"), Data, field, Expression.Constant(args, typeof(BsonValue)));
+		}
 		static bool GTE(BsonDocument document, string name, BsonValue value)
 		{
 			bool ok = false;
@@ -74,6 +89,25 @@ namespace Mdbc
 					return true;
 			}
 			return ok ? false : value.BsonType == BsonType.Null;
+		}
+		static Expression NEExpression(Expression field, BsonValue args)
+		{
+			return Expression.Call(GetMethod("NE"), Data, field, Expression.Constant(args, typeof(BsonValue)));
+		}
+		static bool NE(BsonDocument document, string name, BsonValue value)
+		{
+			bool ok = false;
+			foreach (var data in document.GetNestedValues(name))
+			{
+				ok = true;
+				if (data.CompareTo(value) != 0)
+					return true;
+			}
+			return ok ? false : value.BsonType != BsonType.Null;
+		}
+		static Expression LTExpression(Expression field, BsonValue args)
+		{
+			return Expression.Call(GetMethod("LT"), Data, field, Expression.Constant(args, typeof(BsonValue)));
 		}
 		static bool LT(BsonDocument document, string name, BsonValue value)
 		{
@@ -86,6 +120,10 @@ namespace Mdbc
 			}
 			return ok ? false : value.BsonType == BsonType.Null;
 		}
+		static Expression LTEExpression(Expression field, BsonValue args)
+		{
+			return Expression.Call(GetMethod("LTE"), Data, field, Expression.Constant(args, typeof(BsonValue)));
+		}
 		static bool LTE(BsonDocument document, string name, BsonValue value)
 		{
 			bool ok = false;
@@ -96,6 +134,11 @@ namespace Mdbc
 					return true;
 			}
 			return ok ? false : value.BsonType == BsonType.Null;
+		}
+		static Expression MatchesExpression(Expression field, BsonValue args)
+		{
+			var regex = args.BsonType == BsonType.RegularExpression ? args.AsRegex : null;
+			return Expression.Call(GetMethod("Matches"), Data, field, Expression.Constant(regex, typeof(Regex)));
 		}
 		static bool Matches(BsonDocument document, string name, Regex regex)
 		{
@@ -108,24 +151,48 @@ namespace Mdbc
 
 			return false;
 		}
-		// If the field is missing or the size is empty then false.
-		static bool All(BsonDocument document, string name, BsonArray value)
+		static Expression AllExpression(Expression field, BsonValue args)
 		{
-			if (value.Count == 0)
+			if (args.BsonType != BsonType.Array)
+				throw new ArgumentException("$all argument must be array.");
+
+			var all = new List<object>();
+			foreach (var one in args.AsBsonArray)
+			{
+				//_131116_140311 As Mongo, just check the first element, ignore others
+				BsonDocument d; BsonElement e;
+				if (one.BsonType == BsonType.Document && (d = one.AsBsonDocument).ElementCount > 0 && (e = d.GetElement(0)).Name == "$elemMatch")
+				{
+					if (e.Value.BsonType != BsonType.Document)
+						throw new ArgumentException("$all $elemMatch argument must be document.");
+
+					all.Add(GetFunction(GetExpression(e.Value.AsBsonDocument)));
+				}
+				else
+				{
+					all.Add(one);
+				}
+			}
+			return Expression.Call(GetMethod("All"), Data, field, Expression.Constant(all, typeof(List<object>)));
+		}
+		// If the field is missing or the size is empty then false.
+		static bool All(BsonDocument document, string name, List<object> all)
+		{
+			if (all.Count == 0)
 				return false;
 
 			foreach (var data in document.GetNestedValues(name))
 			{
 				if (data.BsonType != BsonType.Array)
 				{
-					foreach (var it in value)
-						if (data.CompareTo(it) != 0)
+					foreach (var one in all)
+						if (!data.EqualsOrElemMatches(one))
 							goto next;
 					return true;
 				}
 
-				foreach (var it in value)
-					if (data.AsBsonArray.FirstOrDefault(x => it.CompareTo(x) == 0) == null)
+				foreach (var one in all)
+					if (data.AsBsonArray.FirstOrDefault(x => x.EqualsOrElemMatches(one)) == null)
 						goto next;
 				return true;
 
@@ -134,46 +201,72 @@ namespace Mdbc
 
 			return false;
 		}
+		static Expression InExpression(Expression field, BsonValue args)
+		{
+			if (args.BsonType != BsonType.Array)
+				throw new ArgumentException("$in/$nin argument must be array.");
+
+			return Expression.Call(GetMethod("In"), Data, field, Expression.Constant(args.AsBsonArray, typeof(BsonArray)));
+		}
 		static bool In(BsonDocument document, string name, BsonArray value)
 		{
 			foreach (var data in document.GetNestedValues(name))
 			{
 				if (data.BsonType != BsonType.Array)
-					if (value.FirstOrDefault(x => data.CompareTo(x) == 0) != null)
+					if (value.FirstOrDefault(x => data.EqualsOrMatches(x)) != null)
 						return true;
 					else
 						continue;
 
 				foreach (var it in data.AsBsonArray)
-					if (value.FirstOrDefault(x => it.CompareTo(x) == 0) != null)
+					if (value.FirstOrDefault(x => it.EqualsOrMatches(x)) != null)
 						return true;
 			}
 
 			return false;
 		}
+		static Expression ModExpression(Expression field, BsonValue args)
+		{
+			if (args.BsonType != BsonType.Array)
+				throw new ArgumentException("$mod argument must be array.");
+
+			var mod = args.AsBsonArray;
+
+			long mod0 = mod.Count > 0 ? (long)mod[0].ToDoubleOrZero() : 0;
+			if (mod0 == 0)
+				throw new ArgumentException("$mod divisor cannot be 0.");
+
+			long mod1 = mod.Count > 1 ? (long)mod[1].ToDoubleOrZero() : 0;
+
+			return Expression.Call(GetMethod("Mod"), Data, field, Expression.Constant(mod0, typeof(long)), Expression.Constant(mod1, typeof(long)));
+		}
 		// If missing or not a number then false else numbers are converted to long.
-		static bool Mod(BsonDocument document, string name, long modulus, long value)
+		static bool Mod(BsonDocument document, string name, long divisor, long value)
 		{
 			foreach (var data in document.GetNestedValues(name))
 			{
 				switch (data.BsonType)
 				{
 					case BsonType.Int64:
-						if (data.AsInt64 % modulus == value)
+						if (data.AsInt64 % divisor == value)
 							return true;
 						break;
 					case BsonType.Int32:
-						if ((long)data.AsInt32 % modulus == value)
+						if ((long)data.AsInt32 % divisor == value)
 							return true;
 						break;
 					case BsonType.Double:
-						if ((long)data.AsDouble % modulus == value)
+						if ((long)data.AsDouble % divisor == value)
 							return true;
 						break;
 				}
 			}
 
 			return false;
+		}
+		static Expression SizeExpression(Expression field, BsonValue args)
+		{
+			return Expression.Call(GetMethod("Size"), Data, field, Expression.Constant(args.ToDoubleOrZero(), typeof(double)));
 		}
 		// If missing or not a number then false.
 		// If value is not a number then size 0 is used instead.
@@ -186,6 +279,13 @@ namespace Mdbc
 
 			return false;
 		}
+		static Expression TypeExpression(Expression field, BsonValue args)
+		{
+			if (!args.IsNumeric)
+				throw new ArgumentException("$type argument must be number.");
+
+			return Expression.Call(GetMethod("Type"), Data, field, Expression.Constant((BsonType)args.ToInt32(), typeof(BsonType)));
+		}
 		static bool Type(BsonDocument document, string name, BsonType type)
 		{
 			foreach (var data in document.GetNestedValues(name))
@@ -193,6 +293,13 @@ namespace Mdbc
 					return true;
 
 			return false;
+		}
+		static Expression ElemMatchExpression(Expression field, BsonValue args)
+		{
+			if (args.BsonType != BsonType.Document)
+				throw new ArgumentException("$elemMatch argument must be document.");
+
+			return Expression.Call(GetMethod("ElemMatch"), Data, field, Expression.Constant(GetFunction(GetExpression(args.AsBsonDocument)), typeof(Func<BsonDocument, bool>)));
 		}
 		static bool ElemMatch(BsonDocument document, string name, Func<BsonDocument, bool> predicate)
 		{
@@ -214,67 +321,31 @@ namespace Mdbc
 		{
 			return typeof(QueryCompiler).GetMethod(name, BindingFlags.NonPublic | BindingFlags.Static);
 		}
-		static Expression FieldOperatorExpression(string fieldName, string operatorName, BsonValue args)
+		static Expression FieldOperatorExpression(Expression field, string operatorName, BsonValue args)
 		{
-			var field = Expression.Constant(fieldName, typeof(string));
 			switch (operatorName)
 			{
-				case "$exists":
-					var existsValue = LanguagePrimitives.IsTrue(Actor.ToObject(args));
-					return Expression.Equal(
-						Expression.Call(Data, typeof(BsonDocument).GetMethod("Contains"), field),
-						Expression.Constant(existsValue, typeof(bool)));
-				case "$gt":
-					return Expression.Call(GetMethod("GT"), Data, field, Expression.Constant(args, typeof(BsonValue)));
-				case "$gte":
-					return Expression.Call(GetMethod("GTE"), Data, field, Expression.Constant(args, typeof(BsonValue)));
-				case "$lt":
-					return Expression.Call(GetMethod("LT"), Data, field, Expression.Constant(args, typeof(BsonValue)));
-				case "$lte":
-					return Expression.Call(GetMethod("LTE"), Data, field, Expression.Constant(args, typeof(BsonValue)));
-				case "$all":
-					return Expression.Call(GetMethod("All"), Data, field, Expression.Constant(args.AsBsonArray, typeof(BsonArray)));
-				case "$in":
-					return Expression.Call(GetMethod("In"), Data, field, Expression.Constant(args.AsBsonArray, typeof(BsonArray)));
-				case "$nin":
-					return Expression.Not(
-						Expression.Call(GetMethod("In"), Data, field, Expression.Constant(args.AsBsonArray, typeof(BsonArray))));
-				case "$mod":
-					var mod = (BsonArray)args;
-					long mod0 = (long)LanguagePrimitives.ConvertTo(Actor.ToObject(mod[0]), typeof(long), null);
-					long mod1 = (long)LanguagePrimitives.ConvertTo(Actor.ToObject(mod[1]), typeof(long), null);
-					return Expression.Call(GetMethod("Mod"), Data, field,
-						Expression.Constant(mod0, typeof(long)), Expression.Constant(mod1, typeof(long)));
-				case "$ne":
-					return Expression.Call(GetMethod("NE"), Data, field, Expression.Constant(args, typeof(BsonValue)));
-				case "$not":
-					if (args.BsonType == BsonType.Document)
-					{
-						var not = args.AsBsonDocument.GetElement(0);
-						return Expression.Not(FieldOperatorExpression(fieldName, not.Name, not.Value));
-					}
-					else
-					{
-						return Expression.Not(FieldValueExpression(fieldName, args));
-					}
-				case "$regex":
-					var regex = args.BsonType == BsonType.RegularExpression ? args.AsRegex : null;
-					return Expression.Call(GetMethod("Matches"), Data, field, Expression.Constant(regex, typeof(Regex)));
-				case "$size":
-					return Expression.Call(GetMethod("Size"), Data, field, Expression.Constant(args.ToDoubleOrZero(), typeof(double)));
-				case "$type":
-					return Expression.Call(GetMethod("Type"), Data, field, Expression.Constant((BsonType)args.ToInt32(), typeof(BsonType)));
-				case "$elemMatch":
-					var func = GetFunction(GetExpression(args.AsBsonDocument));
-					return Expression.Call(
-						GetMethod("ElemMatch"), Data, field, Expression.Constant(func, typeof(Func<BsonDocument, bool>)));
+				case "$all": return AllExpression(field, args);
+				case "$elemMatch": return ElemMatchExpression(field, args);
+				case "$exists": return ExistsExpression(field, args);
+				case "$gt": return GTExpression(field, args);
+				case "$gte": return GTEExpression(field, args);
+				case "$in": return InExpression(field, args);
+				case "$lt": return LTExpression(field, args);
+				case "$lte": return LTEExpression(field, args);
+				case "$mod": return ModExpression(field, args);
+				case "$ne": return NEExpression(field, args);
+				case "$nin": return Expression.Not(InExpression(field, args));
+				case "$not": return NotExpression(field, args);
+				case "$regex": return MatchesExpression(field, args);
+				case "$size": return SizeExpression(field, args);
+				case "$type": return TypeExpression(field, args);
 				default:
 					throw new NotImplementedException("Not implemented operator " + operatorName);
 			}
 		}
-		static Expression FieldValueExpression(string fieldName, BsonValue value)
+		static Expression FieldValueExpression(Expression field, BsonValue value)
 		{
-			var field = Expression.Constant(fieldName, typeof(string));
 			if (value.IsBsonRegularExpression)
 				return Expression.Call(GetMethod("Matches"), Data, field, Expression.Constant(value.AsRegex, typeof(Regex)));
 			else
@@ -290,16 +361,12 @@ namespace Mdbc
 					{
 						var a = (BsonArray)args;
 						var r = GetExpression((BsonDocument)a[0]);
+
 						for (int i = 1; i < a.Count; ++i)
-							switch (operatorName)
-							{
-								case "$and":
-									r = Expression.And(r, GetExpression((BsonDocument)a[i]));
-									break;
-								default:
-									r = Expression.Or(r, GetExpression((BsonDocument)a[i]));
-									break;
-							}
+						{
+							var e = GetExpression((BsonDocument)a[i]);
+							r = operatorName == "$and" ? Expression.And(r, e) : Expression.Or(r, e);
+						}
 
 						if (operatorName == "$nor")
 							r = Expression.Not(r);
@@ -310,7 +377,7 @@ namespace Mdbc
 					throw new NotImplementedException("Not implemented operator " + operatorName);
 			}
 		}
-		static Expression FieldExpression(string fieldName, BsonValue selector)
+		static Expression FieldExpression(Expression field, BsonValue selector)
 		{
 			var document = selector as BsonDocument;
 			if (document != null && document.ElementCount > 0)
@@ -320,24 +387,24 @@ namespace Mdbc
 				if (operatorName[0] == '$' && operatorName != "$ref")
 				{
 					// combined And on a field: { field: { $ne: 1, $exists: true } }
-					var r = FieldOperatorExpression(fieldName, operatorName, element.Value);
+					var r = FieldOperatorExpression(field, operatorName, element.Value);
 					for (int i = 1; i < document.ElementCount; ++i)
 					{
 						element = document.GetElement(i);
-						r = Expression.And(r, FieldOperatorExpression(fieldName, element.Name, element.Value));
+						r = Expression.And(r, FieldOperatorExpression(field, element.Name, element.Value));
 					}
 					return r;
 				}
 			}
 
-			return FieldValueExpression(fieldName, selector);
+			return FieldValueExpression(field, selector);
 		}
 		static Expression ElementExpression(BsonElement element)
 		{
 			if (element.Name[0] == '$')
 				return OperatorExpression(element.Name, element.Value);
 			else
-				return FieldExpression(element.Name, element.Value);
+				return FieldExpression(Expression.Constant(element.Name, typeof(string)), element.Value);
 		}
 		static Expression DocumentExpression(BsonDocument query)
 		{